JRE Memory Calculator
Previously the values for the applications memory regions were
calculated once during staging.  If the application was scaled such
that the amount of memory allocated to it changed, the application had
to be restaged in order to recalculate how the memory was allocated.
Customers claimed that this slowed down development cycles and the
behavior should be modified such that only a restart was required.

This change incorporates a native executable to calculate the memory
region allocation.  This executable is run during staging to ensure
that the initial memory size and heuristics are valid.  The executable
is then run as a subshell each time the application is started.  This
means that  when the memory size is changed and the application is
restarted, the memory allocations are adjusted accordingly.  If the
new memory size causes an illegal allocation of memory, startup will
fail.

[#93606820]

+++ spec/application_helper.rb
@@ -65,8 +65,12 @@ shared_context 'application_helper' do
     application
   end
 
-  after do
-    FileUtils.rm_rf app_dir
+  after do |example|
+    if example.metadata[:no_cleanup]
+      puts "Application Directory: #{app_dir}"
+    else
+      FileUtils.rm_rf app_dir
+    end
   end
 
 end
