+++ spec/integration_helper.rb
@@ -38,8 +38,12 @@ shared_context 'integration_helper' do
     FileUtils.cp_r "spec/fixtures/#{buildpack_fixture.chomp}/.", buildpack_dir if buildpack_fixture
   end
 
-  after do
-    FileUtils.rm_rf buildpack_dir
+  after do |example|
+    if example.metadata[:no_cleanup]
+      puts "Buildpack Directory: #{buildpack_dir}"
+    else
+      FileUtils.rm_rf buildpack_dir
+    end
   end
 
   def run(command)
